bot: report: adjust profile_summary
Profile_summary function is split to two functions that can return
ascii formatted table or html formatted. Html version printed in readme.md
was incorrectly formatted by github.
